// generated/src/app/api-models/batch-master-out.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class BatchMasterOut {
  batchId: string;
  totalTransactions: number;
  sourceTpId: string;
  description: string;
  claimType: string;
  filterText1: string;
  filterFlag1: string;
  filterText2: string;
  filterFlag2: string;
  filterText3: string;
  filterFlag3: string;
  filterText4: string;
  filterFlag4: string;
  filterText5: string;
  filterFlag5: string;
  filterText6: string;
  filterFlag6: string;
  filterText7: string;
  filterFlag7: string;
  filterText8: string;
  filterFlag8: string;
  filterText9: string;
  filterFlag9: string;
  filterText10: string;
  filterFlag10: string;
  filterDate1: Date;
  filterDate2: Date;
  filterDate3: Date;
  filterDate4: Date;
  filterDate5: Date;
  filterDate6: Date;
  filterDate7: Date;
  filterDate8: Date;
  filterDate9: Date;
  filterDate10: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  filterText11: string;
  filterFlag11: string;
  filterText12: string;
  filterFlag12: string;
  filterText13: string;
  filterFlag13: string;
  filterText14: string;
  filterFlag14: string;
  filterText15: string;
  filterFlag15: string;
  filterText16: string;
  filterFlag16: string;
}